Break out executor into file, clean up methods
Draining was happening out of order with stopping the
container watcher, which also caused some weird behavior
in testing.
